--- a/include/simdjson/generic/ondemand/token_position.h
+++ b/include/simdjson/generic/ondemand/token_position.h
@@ -1,0 +1,10 @@
+namespace simdjson {
+namespace SIMDJSON_IMPLEMENTATION {
+namespace ondemand {
+
+/** @private Position in the JSON buffer indexes */
+using token_position = const uint32_t *;
+
+} // namespace ondemand
+} // namespace SIMDJSON_IMPLEMENTATION
+} // namespace simdjson