This PR makes it so you can keep a `value` around as long as you like (even after the document is parsed), and still convert it to a scalar (string, number, boolean, null) correctly and without error.

It also removes the distinction between `&&` and `&` when retrieving scalar values: when you first attempt to parse a value, the iterator is *always* advanced, even if it fails. Subsequent attempts (for example, trying other types) will reach back to read the value.

The biggest reason to do this, besides the apparent performance gain on existing benchmarks, is that it makes some constructs work that would be confusing otherwise. The principal one is one I wrote myself: `std::pair<double, double>(tweet["x"], tweet["y"])` works correctly with this patch. Another scenario this fixes:

```c++
for (auto tweet : doc["statuses"]) {
  simdjson::value name;
  if (tweet["user"]["verified"]) {
    name = tweet["user"]["name"];
  } else {
    name = tweet["user"]["screen_name"];
  }
  std::cout << "Name: " << name << std::endl;
}
```

Silly scenario for twitter, but I'm sure we've both seen records where the data you want is either in field X or Y depending on the value in field Z. There are other ways to write any of these, but as they do compile, it's definitely better that they *work*.

### Top Tweet Scenario

For an example scenario, consider finding the "top tweet"--the tweet with the highest retweet_count. Since you can't know which is the top tweet until you scan them all, you can't know which tweet you're going to need to parse into a result. Therefore, a purely forward-only API with no lazy value support will be forced to parse each candidate tweet just in case it turns out to be the top tweet.

```c++
// Get the screen name and text of the tweet with the most retweets
uint64_t max_retweets;
ondemand::value top_screen_name, top_text;
for (auto tweet : doc["statuses"]) {
  uint64_t retweets = tweet["retweet_count"];
  if (retweets > max_retweets) {
    top_text = tweet["text"];
    top_screen_name = tweet["user"]["screen_name"];
  }
}
std::cout << "Most retweeted tweet (with " << max_retweets << " tweets) was by user " << std::string_view(top_screen_name) << std::endl;
std::cout << std::string_view(top_text) << std::endl;
```

By delaying the (relatively expensive) parsing of text until the end, you can do just *one* screen name and text instead of having to extract every one.

#### Top Tweet Scenario Performance (Skylake Clang 10)

I implemented two versions of the above benchmark:

* `ondemand` saves the text and screen_name JSON scalar values without parsing. This means we repeat the *lookup* over and over, but not the parse.
* `ondemand_forwardonly` parses text and screen_name every time, to highlight the savings of this feature in particular.

ondemand_forwardonly is *faster*. I cannot tell you why. It is still absolutely true that parsing values later does less work, and I still want this checkin due to the fact that it makes some stuff work that didn't before ... but we should make sure to come back to this one day :)

```
----------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------
Benchmark                                                      Time             CPU   Iterations best_branch_miss best_bytes_per_sec best_cache_miss best_cache_ref best_cycles best_cycles_per_byte best_docs_per_sec best_frequency best_instructions best_instructions_per_byte best_instructions_per_cycle best_items_per_sec branch_miss      bytes bytes_per_second cache_miss  cache_ref     cycles cycles_per_byte docs_per_sec  frequency instructions instructions_per_byte instructions_per_cycle      items items_per_second
----------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------
top_tweet<simdjson_dom>/manual_time                       258876 ns       287022 ns         2704           3.261k           2.45534G               0        86.454k    949.692k              1.50383          3.88801k       3.69241G          2.88216M                    4.56388                     3.03484           3.88801k    3.43207k   631.515k       2.27192G/s   0.768861   86.4597k   955.407k         1.51288   3.86286k/s  3.6906G/s     2.88216M               4.56388                3.01668          1       3.86286k/s [BEST: throughput=  2.46 GB/s doc_throughput=  3888 docs/s instructions=     2882159 cycles=      949692 branch_miss=    3261 cache_miss=       0 cache_ref=     86454 items=         1 avg_time=    258875 ns]
top_tweet<simdjson_ondemand>/manual_time                  149617 ns       177518 ns         4672           1.513k           4.25079G               0        52.936k    548.711k              0.86888          6.73111k       3.69343G          1.91575M                    3.03358                     3.49137           6.73111k    1.62116k   631.515k       3.93101G/s  0.0432363   52.9805k   552.307k        0.874575   6.68375k/s 3.69148G/s     1.91575M               3.03358                3.46863          1       6.68375k/s [BEST: throughput=  4.25 GB/s doc_throughput=  6731 docs/s instructions=     1915751 cycles=      548711 branch_miss=    1513 cache_miss=       0 cache_ref=     52936 items=         1 avg_time=    149616 ns]
top_tweet<simdjson_ondemand_forward_only>/manual_time     160857 ns       188864 ns         4353           1.463k           3.97239G               0        53.851k    587.123k             0.929706          6.29026k       3.69315G          1.92366M                    3.04611                     3.27642           6.29026k    1.61333k   631.515k       3.65633G/s  0.0259591   53.8618k   593.784k        0.940253   6.21672k/s 3.69139G/s     1.92366M               3.04611                3.23967          1       6.21672k/s [BEST: throughput=  3.97 GB/s doc_throughput=  6290 docs/s instructions=     1923664 cycles=      587123 branch_miss=    1463 cache_miss=       0 cache_ref=     53851 items=         1 avg_time=    160856 ns]
top_tweet<yyjson>/manual_time                             625825 ns       655530 ns         1125          10.141k            1.7421G             618        69.202k     1.2998M              2.05823          2.75861k       3.58564G          2.82777M                    4.47776                     2.17554           2.75861k    7.78023k   631.515k       962.345M/s    1020.73   44.5412k   1.51008M         2.39121   1.59789k/s 2.41295G/s     2.82835M               4.47868                1.87298          1       1.59789k/s [BEST: throughput=  1.74 GB/s doc_throughput=  2758 docs/s instructions=     2827770 cycles=     1299801 branch_miss=   10141 cache_miss=     618 cache_ref=     69202 items=         1 avg_time=    625825 ns]
top_tweet<yyjson_insitu>/manual_time                      425042 ns       454211 ns         1647           7.133k           2.04327G             312        40.429k    1.13499M              1.79725           3.2355k       3.67226G          2.82775M                    4.47772                     2.49143            3.2355k    7.57587k   631.515k       1.38373G/s    48.4323   26.9132k   1.23679M         1.95844   2.35271k/s 2.90979G/s     2.82801M               4.47814                2.28658          1       2.35271k/s [BEST: throughput=  2.04 GB/s doc_throughput=  3235 docs/s instructions=     2827746 cycles=     1134990 branch_miss=    7133 cache_miss=     312 cache_ref=     40429 items=         1 avg_time=    425042 ns]
top_tweet<sajson>/manual_time                             582843 ns       611379 ns         1201          10.398k           1094.49M               0        41.191k    2.13017M              3.37311          1.73312k       3.69184G          5.74294M                    9.09391                       2.696           1.73312k    10.7316k   631.515k       1033.31M/s   0.843464   41.3345k   2.15073M         3.40566   1.71573k/s 3.69006G/s     5.74294M               9.09391                2.67023          1       1.71573k/s [BEST: throughput=  1.09 GB/s doc_throughput=  1733 docs/s instructions=     5742943 cycles=     2130168 branch_miss=   10398 cache_miss=       0 cache_ref=     41191 items=         1 avg_time=    582843 ns]
top_tweet<rapidjson>/manual_time                         2155200 ns      2184293 ns          325          32.404k           321.623M         14.557k        50.628k    7.24841M              11.4778           509.288       3.69153G          21.8021M                    34.5235                     3.00784            509.288    32.6587k   631.515k       279.445M/s    478.412   29.8018k   7.38989M         11.7018    463.994/s 3.42887G/s      21.803M                34.525                2.95039          1        463.994/s [BEST: throughput=  0.32 GB/s doc_throughput=   509 docs/s instructions=    21802077 cycles=     7248406 branch_miss=   32404 cache_miss=   14557 cache_ref=     50628 items=         1 avg_time=   2155199 ns]
top_tweet<rapidjson_insitu>/manual_time                  1513668 ns      1542722 ns          466          25.438k           450.833M              62        39.181k    5.17121M              8.18857           713.892       3.69168G          13.2241M                    20.9403                     2.55726            713.892    26.4094k   631.515k       397.881M/s    218.444   29.9593k   5.25642M         8.32351    660.647/s 3.47264G/s     13.2245M               20.9409                2.51587          1        660.647/s [BEST: throughput=  0.45 GB/s doc_throughput=   713 docs/s instructions=    13224096 cycles=     5171207 branch_miss=   25438 cache_miss=      62 cache_ref=     39181 items=         1 avg_time=   1513667 ns]
top_tweet<nlohmann_json>/manual_time                     8909969 ns      8940213 ns           79         196.661k           71.2889M              10       377.498k    32.6837M              51.7544           112.886       3.68952G          85.4176M                    135.258                     2.61346            112.886    200.142k   631.515k       67.5939M/s    13.8228   380.493k   32.8712M         52.0513    112.234/s 3.68926G/s     85.7546M               135.792                2.60881          1        112.234/s [BEST: throughput=  0.07 GB/s doc_throughput=   112 docs/s instructions=    85417566 cycles=    32683690 branch_miss=  196661 cache_miss=      10 cache_ref=    377498 items=         1 avg_time=   8909968 ns]
```

### Other Things It Solves

This also solves the issue I had originally with `make_pair<double, double>(obj["x"], obj["y"])`, correctly turning x and y into a double even though that doesn't happen until the iterator is sitting at "y".

This does *not* attempt a similar thing with arrays or objects. Those will require more design than this.

## Performance vs. master

A *key* part of this is that in-order parsing should not get even a little bit slower. We accomplish this with an if statement when we get the JSON text for the scalar we want to parse:

```c++
if (!is_at_start()) { return iter.peek(_start_index); } // don't advance the iterator if we're not using it
else { return iter.advance(); } // *do* advance the iterator if we're using the iterator
```

This way, when it's used in order, the iterator is properly advanced when the value is used. When it's used out of order, it just reaches back to find the value and doesn't worry about the iterator.

My belief is that this is optimized away for single-use values (such as `double x = obj["x"]`). This is because, when the `value_iterator` is created, its _start_index is set to `json_iterator->token.index`. Then the very next thing that's done is `if (_start_index == json_iterator->token.index) {` ... which the compiler can optimize away. Not only that, it knows the value of start_index is already stored in token.index (which should be a register already), so it doesn't need to allocate a register or memory to it--the _start_index *itself* should be optimized away.

### Skylake (Clang 10)

Performance numbers seem to bear this out, showing no downside, and a little bit of *upside* on partial_tweets and find_tweet:

PR:

```
----------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------
Benchmark                                                      Time             CPU   Iterations best_branch_miss best_bytes_per_sec best_cache_miss best_cache_ref best_cycles best_cycles_per_byte best_docs_per_sec best_frequency best_instructions best_instructions_per_byte best_instructions_per_cycle best_items_per_sec branch_miss      bytes bytes_per_second cache_miss  cache_ref     cycles cycles_per_byte docs_per_sec  frequency instructions instructions_per_byte instructions_per_cycle      items items_per_second
----------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------
partial_tweets<simdjson_ondemand>/manual_time             155736 ns       183895 ns         4496           1.426k           4.08586G               0        54.665k    570.869k             0.903967          6.46994k       3.69349G          1.94927M                    3.08666                     3.41457           646.994k    1.56714k   631.515k       3.77654G/s  0.0531584   54.8954k   574.893k         0.91034   6.42112k/s 3.69146G/s     1.94927M               3.08666                3.39067        100       642.112k/s [BEST: throughput=  4.09 GB/s doc_throughput=  6469 docs/s instructions=     1949274 cycles=      570869 branch_miss=    1426 cache_miss=       0 cache_ref=     54665 items=       100 avg_time=    155736 ns]
Creating a source file spanning 44921 KB 
large_random<simdjson_ondemand>/manual_time             66933147 ns     70347823 ns           10         908.667k           687.718M        5.77398M       7.90181M    246.763M              5.36455           14.9508        3.6893G          689.848M                    14.9971                     2.79559           14.9508M    909.785k   45.9988M       655.398M/s   5.77303M   7.89844M   246.922M         5.36801    14.9403/s 3.68908G/s     689.848M               14.9971                2.79379      1000k       14.9403M/s [BEST: throughput=  0.69 GB/s doc_throughput=    14 docs/s instructions=   689847657 cycles=   246762771 branch_miss=  908667 cache_miss= 5773977 cache_ref=   7901813 items=   1000000 avg_time=  66933147 ns]
large_random<simdjson_ondemand_unordered>/manual_time   66587053 ns     69998774 ns           11         915.819k           691.137M        5.75063M       7.89995M    245.541M              5.33798           15.0251       3.68928G          690.848M                    15.0188                     2.81358           15.0251M    915.218k   45.9988M       658.805M/s   5.76904M   7.90159M   245.646M         5.34027    15.0179/s 3.68909G/s     690.848M               15.0188                2.81237      1000k       15.0179M/s [BEST: throughput=  0.69 GB/s doc_throughput=    15 docs/s instructions=   690847915 cycles=   245540733 branch_miss=  915819 cache_miss= 5750632 cache_ref=   7899946 items=   1000000 avg_time=  66587053 ns]
Creating a source file spanning 134087 KB 
kostya<simdjson_ondemand>/manual_time                   60349654 ns     70747151 ns           12          459.96k           2.27674G        10.0868M       13.9045M    222.481M              1.62034           16.5816       3.68908G          660.474M                    4.81027                     2.96868           8.69353M    459.832k   137.305M       2.11891G/s   10.0962M   13.8133M   222.634M         1.62146    16.5701/s 3.68907G/s     660.474M               4.81027                2.96663   524.288k       8.68751M/s [BEST: throughput=  2.28 GB/s doc_throughput=    16 docs/s instructions=   660473985 cycles=   222480501 branch_miss=  459960 cache_miss=10086798 cache_ref=  13904489 items=    524288 avg_time=  60349654 ns]
distinct_user_id<simdjson_ondemand>/manual_time           149454 ns       178779 ns         4676           1.555k           4.28614G               0        52.782k    544.147k             0.861653          6.78707k       3.69316G          1.89752M                    3.00471                     3.48714           780.513k    1.80153k   631.515k       3.93529G/s  0.0530368   52.7491k   551.693k        0.873603   6.69103k/s 3.69139G/s     1.89752M               3.00471                3.43944        115       769.468k/s [BEST: throughput=  4.29 GB/s doc_throughput=  6787 docs/s instructions=     1897519 cycles=      544147 branch_miss=    1555 cache_miss=       0 cache_ref=     52782 items=       115 avg_time=    149453 ns]
find_tweet<simdjson_ondemand>/manual_time                 105333 ns       133367 ns         6626              629           6.02039G               0        30.093k    387.529k              0.61365          9.53325k       3.69441G          1.39043M                    2.20173                     3.58793           9.53325k     655.537   631.515k       5.58368G/s   9.96076m    30.186k   388.885k        0.615797   9.49372k/s 3.69197G/s     1.39043M               2.20173                3.57542          1       9.49372k/s [BEST: throughput=  6.02 GB/s doc_throughput=  9533 docs/s instructions=     1390428 cycles=      387529 branch_miss=     629 cache_miss=       0 cache_ref=     30093 items=         1 avg_time=    105332 ns]
```

master:

```
----------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------
…