@jkeiser 2021-01-10
@lemire I'd love to release 0.8 with this in it, but I think we'll need to still call it a preview: the remaining changes (particularly the ability to use arrays and objects lazily) may require some I think there are potential name changes before we call it a beta.

Recursion support will impact things pretty drastically too, I think (how drastically remains to be seen), but I think we can release a 1.0 without it. We can't replace DOM until we have it, but I'm OK with not deprecating DOM in 1.0. Release early, release often.