@jkeiser 2021-01-10
Added a benchmark with a new "most retweeted tweet" scenario designed to benchmark an implementation's ability to lazily parse. I think it's real-world enough to include as a benchmark. The basic idea is, you find the tweet with the most retweets, and output the screen_name and text of that tweet. Truly forward-only implementations (or non-lazy DOMs) will have to parse a lot screen_names and texts, because they don't know *which* one they will need until they've gone through them all. Lazy implementations can just update a "top_tweet" value pointing at the tweet object, and parse the stuff they need later.