@jkeiser 2021-01-10
In a fascinating turn of events I don't entirely understand, parsing the text and screen name of most of the entries (forward-only) is cheaper than saving the text and screen_name fields separately, but is *not* cheaper than saving the entire tweet object.

I've realized this doesn't have to be limited to scalar values, so I'm going to lift that restriction. The caveat is that we'll need a safety guardrail in there to ensure that you can only perform *nested* iteration, and not *parallel* iteration.