@jkeiser 2021-01-10
Finally, I've discovered you can't do `value a; simdjson_result<value> b = obj["x"]; a = b;` as it is basically assuming these are move-only objects (they are not). I will fix that here.