@jkeiser 2021-01-10
Never mind, the assignment problem is a separate one. It sucks you have std::move() sometimes, but that's a beta for ya.