@jkeiser 2021-01-10
Arrays and objects are separate too. The lazy object worked, but it was a total fluke due to how the internals work. Objects and arrays still need their own PR. Possibly after 0.8.