@jkeiser 2021-01-10
Since I added new benchmarks, I added sajson and nlohmann_json to them, and I'm setting this PR to merge into #1372.