@jkeiser 2021-01-13
What was weird about it?