@lemire 2021-01-14
@jkeiser The weird test failure. It was complaining about a missing branch.

 We can now merge! Only travis is failing...